create new classes to shrink TaxonomyController and avoid anti-pattern of running db queries in the controller
