+++ app/Models/TaxaResourceLink.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class TaxaResourceLink extends Model {
+    protected $table = 'taxaresourcelinks';
+
+    public $timestamps = false;
+}
