+++ app/Models/TaxaVernacular.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class TaxaVernacular extends Model {
+    protected $table = 'taxavernaculars';
+
+    public $timestamps = false;
+}
